남현수 - Employee 생성, 수정, 삭제하면 change_log 생성 및 histories 생성

+++ hr_bank/src/main/java/com/codeit/hrbank/service/impl/EmployeeServiceImpl.java
@@ -5,18 +5,25 @@ import com.codeit.hrbank.dto.request.EmployeeCreateRequest;
 import com.codeit.hrbank.dto.request.EmployeeUpdateRequest;
 import com.codeit.hrbank.dto.response.CursorPageResponse;
 import com.codeit.hrbank.entity.BinaryContent;
+import com.codeit.hrbank.entity.ChangeLog;
+import com.codeit.hrbank.entity.ChangeLogType;
 import com.codeit.hrbank.entity.Department;
 import com.codeit.hrbank.entity.Employee;
 import com.codeit.hrbank.entity.EmployeeStatus;
+import com.codeit.hrbank.entity.History;
 import com.codeit.hrbank.mapper.EmployeeMapper;
 import com.codeit.hrbank.mapper.PageResponseMapper;
 import com.codeit.hrbank.repository.BinaryContentRepository;
+import com.codeit.hrbank.repository.ChangeLogRepository;
 import com.codeit.hrbank.repository.DepartmentRepository;
 import com.codeit.hrbank.repository.EmployeeRepository;
+import com.codeit.hrbank.repository.HistoryRepository;
 import com.codeit.hrbank.service.EmployeeService;
 import com.codeit.hrbank.storage.BinaryContentStorage;
+import com.codeit.hrbank.util.ChangeLogUtils;
+import jakarta.persistence.EntityManager;
 import java.io.IOException;
-import java.time.Instant;
+import java.time.LocalDate;
 import java.time.ZoneId;
 import java.util.List;
 import java.util.NoSuchElementException;
@@ -34,13 +41,17 @@ public class EmployeeServiceImpl implements EmployeeService {
   private final DepartmentRepository departmentRepository;
   private final BinaryContentRepository binaryContentRepository;
   private final BinaryContentStorage binaryContentStorage;
+  private final ChangeLogRepository changeLogRepository;
+  private final HistoryRepository historyRepository;
 
   private final EmployeeMapper employeeMapper;
   private final PageResponseMapper pageResponseMapper;
 
+  private final EntityManager entityManager;
+
   @Override
   @Transactional
-  public EmployeeDTO createEmployee(EmployeeCreateRequest request, MultipartFile profile) {
+  public EmployeeDTO createEmployee(EmployeeCreateRequest request, MultipartFile profile, String ipAddress) {
     if (employeeRepository.existsByEmail(request.email())) {
       throw new IllegalArgumentException("중복되는 이메일이 있습니다.");
     }
@@ -48,16 +59,17 @@ public class EmployeeServiceImpl implements EmployeeService {
     Department department = departmentRepository.findById(request.departmentId())
         .orElseThrow(() -> new NoSuchElementException("부서를 찾을 수 없습니다."));
 
-    // employeeNumber 생성 (EMP-YYYY-UUID)
     int year = request.hireDate().atStartOfDay(ZoneId.systemDefault()).getYear();
     String randomPart = UUID.randomUUID().toString().substring(0, 8).toUpperCase();
-
     String newEmployeeNumber = String.format("EMP-%d-%s", year, randomPart);
 
     BinaryContent binaryContent = null;
     if (profile != null && !profile.isEmpty()) {
-      binaryContent = BinaryContent.builder().name(profile.getOriginalFilename())
-          .size(profile.getSize()).contentType(profile.getContentType()).build();
+      binaryContent = BinaryContent.builder()
+          .name(profile.getOriginalFilename())
+          .size(profile.getSize())
+          .contentType(profile.getContentType())
+          .build();
       binaryContentRepository.save(binaryContent);
       try {
         binaryContentStorage.putFile(binaryContent.getId(), profile.getBytes(), binaryContent.getName());
@@ -65,18 +77,27 @@ public class EmployeeServiceImpl implements EmployeeService {
         throw new RuntimeException(e);
       }
     }
-    // 꼼수 바꿔야하면 바꿔야함.
-    Instant hireDate = request.hireDate().atStartOfDay(ZoneId.systemDefault()).toInstant();
 
-    Employee employee = Employee.builder().name(request.name()).email(request.email())
-        .employeeNumber(newEmployeeNumber).department(department).position(request.position())
-        .hireDate(hireDate).status(EmployeeStatus.ACTIVE).binaryContent(binaryContent)
-        // 메모는 어따가 넣어?
+    Employee employee = Employee.builder()
+        .name(request.name())
+        .email(request.email())
+        .employeeNumber(newEmployeeNumber)
+        .department(department)
+        .position(request.position())
+        .hireDate(request.hireDate())
+        .status(EmployeeStatus.ACTIVE)
+        .binaryContent(binaryContent)
         .build();
+    Employee savedEmployee = employeeRepository.save(employee);
 
-    Employee save = employeeRepository.save(employee);
+    // ChangeLog + History
+    ChangeLog changeLog = ChangeLogUtils.createChangeLog(ChangeLogType.CREATED, ipAddress, request.memo(), savedEmployee);
+    changeLogRepository.save(changeLog);
 
-    return employeeMapper.toDTO(save);
+    List<History> histories = ChangeLogUtils.createHistoriesForCreate(changeLog, savedEmployee);
+    historyRepository.saveAll(histories);
+
+    return employeeMapper.toDTO(savedEmployee);
   }
 
   @Override
@@ -90,8 +111,7 @@ public class EmployeeServiceImpl implements EmployeeService {
 
   @Override
   @Transactional
-  public EmployeeDTO updateEmployee(Long employeeId, EmployeeUpdateRequest request,
-      MultipartFile profile) {
+  public EmployeeDTO updateEmployee(Long employeeId, EmployeeUpdateRequest request, MultipartFile profile, String ipAddress) {
     Employee employee = employeeRepository.findById(employeeId)
         .orElseThrow(() -> new NoSuchElementException("회원이 존재하지 않습니다"));
 
@@ -104,8 +124,11 @@ public class EmployeeServiceImpl implements EmployeeService {
 
     BinaryContent binaryContent = null;
     if (profile != null && !profile.isEmpty()) {
-      binaryContent = BinaryContent.builder().name(profile.getOriginalFilename())
-          .size(profile.getSize()).contentType(profile.getContentType()).build();
+      binaryContent = BinaryContent.builder()
+          .name(profile.getOriginalFilename())
+          .size(profile.getSize())
+          .contentType(profile.getContentType())
+          .build();
       binaryContentRepository.save(binaryContent);
       try {
         binaryContentStorage.putFile(binaryContent.getId(), profile.getBytes(), binaryContent.getName());
@@ -114,43 +137,54 @@ public class EmployeeServiceImpl implements EmployeeService {
       }
     }
 
-    // 꼼수 바꿔야하면 바꿔야함.
-    // 입사일 업데이트: null이면 기존 값 유지
-    Instant hireDate = employee.getHireDate();
-
-    Employee updateEmployee = employee.toBuilder()
+    Employee updatedEmployee = employee.toBuilder()
         .name(request.name() != null ? request.name() : employee.getName())
         .email(request.email() != null ? request.email() : employee.getEmail())
-        .department(department != null ? department : employee.getDepartment())
+        .department(department)
         .position(request.position() != null ? request.position() : employee.getPosition())
-        .hireDate(hireDate != null ? hireDate: employee.getHireDate())
+        .hireDate(employee.getHireDate())
         .binaryContent(binaryContent != null ? binaryContent : employee.getBinaryContent())
-        // 메모는 없으면 기존 메모 유지
         .build();
 
-    Employee save = employeeRepository.save(updateEmployee);
-    return employeeMapper.toDTO(save);
+    Employee savedEmployee = employeeRepository.save(updatedEmployee);
+
+    ChangeLog changeLog = ChangeLogUtils.createChangeLog(ChangeLogType.UPDATED, ipAddress, request.memo(), savedEmployee);
+    changeLogRepository.save(changeLog);
+
+    List<History> histories = ChangeLogUtils.createHistoriesForUpdate(changeLog, employee, savedEmployee);
+    if (!histories.isEmpty()) {
+      historyRepository.saveAll(histories);
+    }
+
+    return employeeMapper.toDTO(savedEmployee);
   }
 
   @Override
   @Transactional
-  public void deleteEmployee(Long employeeId) {
+  public void deleteEmployee(Long employeeId, String ipAddress) {
     Employee employee = employeeRepository.findById(employeeId)
         .orElseThrow(() -> new NoSuchElementException("회원이 존재하지 않습니다."));
 
-    employeeRepository.deleteById(employee.getId());
+    ChangeLog changeLog = ChangeLogUtils.createChangeLog(ChangeLogType.DELETED, ipAddress, null, employee);
+    changeLogRepository.save(changeLog);
 
-    binaryContentRepository.deleteById(employee.getBinaryContent().getId());
-    binaryContentStorage.deleteFile(employee.getBinaryContent().getId());
+    List<History> histories = ChangeLogUtils.createHistoriesForDelete(changeLog, employee);
+    historyRepository.saveAll(histories);
+
+    employeeRepository.deleteById(employeeId);
+
+    if (employee.getBinaryContent() != null) {
+      binaryContentRepository.deleteById(employee.getBinaryContent().getId());
+      binaryContentStorage.deleteFile(employee.getBinaryContent().getId());
+    }
   }
 
   @Override
   @Transactional(readOnly = true)
   public CursorPageResponse<EmployeeDTO> findAllByPart(String nameOrEmail, String employeeNumber,
-      String departmentName, String position, Instant hireDateFrom, Instant hireDateTo,
+      String departmentName, String position, LocalDate hireDateFrom, LocalDate hireDateTo,
       EmployeeStatus status, Long idAfter, Integer size, String sortField, String sortDirection) {
 
-    // 1. Entity 조회 (QueryDSL)
     List<Employee> employees = employeeRepository.findAllQEmployeesPart(
         nameOrEmail, employeeNumber, departmentName, position,
         hireDateFrom, hireDateTo, status, idAfter, size, sortField, sortDirection
@@ -158,23 +192,20 @@ public class EmployeeServiceImpl implements EmployeeService {
 
     boolean hasNext = employees.size() > size;
     if (hasNext) {
-      employees = employees.subList(0, size); // 초과분 잘라내기
+      employees = employees.subList(0, size);
     }
 
-    // 2. Entity → DTO 변환
     List<EmployeeDTO> employeeDTOs = employees.stream()
         .map(employeeMapper::toDTO)
         .toList();
 
-    // 3. 다음 커서 계산
     Long nextIdAfter = null;
     String nextCursor = null;
     if (!employees.isEmpty()) {
       nextIdAfter = employees.get(employees.size() - 1).getId();
-      nextCursor = String.valueOf(nextIdAfter); // cursor 확장 고려 시 Base64 encode 가능
+      nextCursor = String.valueOf(nextIdAfter);
     }
 
-    // 4. CursorPageResponse 생성
     return pageResponseMapper.fromCursor(employeeDTOs, size, nextCursor, nextIdAfter, hasNext);
   }
 }
